% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKGML.R
\name{childIsEntry}
\alias{childIsEntry}
\title{childIsEntry}
\usage{
childIsEntry(r)
}
\arguments{
\item{r}{xml root}
}
\value{
T or F
}
\description{
childIsEntry
}
\examples{
path<-system.file(package = "Key2Enrich")
xmlPath<-paste(path,"/extdata",sep="")
xmlFile<-paste(xmlPath,sep="","/hsa04012.xml")
r<-parseKGMLFile(xmlFile)
kegg.nodes <- lapply(r[childIsEntry(r)], parseEntry)
}
