% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{get_n}
\alias{get_n}
\title{Get the number of sample genes with KEGG gene ID}
\usage{
get_n(inputSampleKEGG, allGeneInPathwayDF)
}
\arguments{
\item{inputSampleKEGG}{input sample with in KEGG gene ID}

\item{allGeneInPathwayDF}{all KEGG pathway ID and its gene list in KEGG gene ID format}
}
\value{
n: the number of sample genes with KEGG gene ID
}
\description{
Get the number of sample genes with KEGG gene ID
}
\examples{
#get_n(inputSampleKEGG,allGeneInPathwayDF)

}
