% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{getAllGeneInPathwayDF}
\alias{getAllGeneInPathwayDF}
\title{Get all KEGG pathway ID and its corresponding gene list in KEGG gene ID format}
\usage{
getAllGeneInPathwayDF(KEGGSpecies)
}
\arguments{
\item{KEGGSpecies}{species in KEGG format}
}
\value{
KEGG pathway ID, and its gene list in KEGG Gene ID format
}
\description{
Get all KEGG pathway ID and its corresponding gene list in KEGG gene ID format
}
\examples{
getAllGeneInPathwayDF("mmu")
}
