% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{getAllPathNameAndID}
\alias{getAllPathNameAndID}
\title{Get all KEGG pathway name and ID of species}
\usage{
getAllPathNameAndID(KEGGSpecies)
}
\arguments{
\item{KEGGSpecies}{species in KEGG format}
}
\value{
KEGG pathway ID of input species
}
\description{
Get all KEGG pathway name and ID of species
}
\examples{
#getAllPathNameAndID("mmu")
}
