% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioMart.R
\name{getAllTypesID}
\alias{getAllTypesID}
\title{Convert input file ID to dataframe class with entrez ID, symbol and ensembl ID}
\usage{
getAllTypesID(inputSpecies, inputIDType, geneIDCol)
}
\arguments{
\item{inputSpecies}{human, mouse, rat}

\item{geneIDCol}{the input ID}

\item{IDType}{Human: entrezgene, hgnc_symbol, ensembl_gene_id, Mouse: entrezgene, mgi_symbol, ensembl_gene_id, Rat: entrezgene, rgd_symbol, ensembl_gene_id}
}
\value{
input sample in dataframe class with entrez ID, symbol and ensembl ID
}
\description{
Convert input file ID to dataframe class with entrez ID, symbol and ensembl ID
}
\examples{
getAllTypesID("mouse","entrezgene",geneIDCol)
}
