% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{getColor}
\alias{getColor}
\title{Get color for node from log fold change value}
\usage{
getColor(thisValue)
}
\arguments{
\item{thisValue}{log fold change value of input}
}
\value{
color for node
}
\description{
Get color for node from log fold change value
}
\examples{
#getColor(1.6)
}
