% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{getGeneInOnePathwayDF}
\alias{getGeneInOnePathwayDF}
\title{Get gene list in specific KEGG pathway}
\usage{
getGeneInOnePathwayDF(thispathID, KEGGSpecies)
}
\arguments{
\item{thispathID}{KEGG pathway ID}

\item{KEGGSpecies}{species in KEGG format}
}
\value{
gene list in specific KEGG pathway
}
\description{
Get gene list in specific KEGG pathway
}
\examples{
#library(Key2Enrich)
#getGeneInOnePathwayDF("mmu05160","mmu")
}
