% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{getGeneSymbol}
\alias{getGeneSymbol}
\title{Get gene symbol from different species}
\usage{
getGeneSymbol(thisKEGGSpecies)
}
\arguments{
\item{thisKEGGSpecies}{species in KEGG format}
}
\value{
name of gene symbol
}
\description{
Get gene symbol from different species
}
