% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{getIndex2}
\alias{getIndex2}
\title{Get index of value in array for left color key}
\usage{
getIndex2(thisArray, thisValue)
}
\arguments{
\item{thisArray}{color array}

\item{thisValue}{log fold change value}
}
\value{
index
}
\description{
Get index of value in array for left color key
}
