% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{getN}
\alias{getN}
\title{Get the number of all genes with KEGG gene ID}
\usage{
getN(allGeneInPathwayDF)
}
\arguments{
\item{allGeneInPathwayDF}{all KEGG pathway ID and its gene list in KEGG gene ID format}
}
\value{
N:the number of all genes with KEGG gene ID
}
\description{
Get the number of all genes with KEGG gene ID
}
\examples{
library(Key2Enrich)
allGeneInPathwayDF<-getAllGeneInPathwayDF("mmu")
print(N<-getN(allGeneInPathwayDF))

}
