% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKGML.R
\name{getPathInfo}
\alias{getPathInfo}
\title{Get KEGG pathway info from xmlRoot}
\usage{
getPathInfo(r)
}
\arguments{
\item{r}{xmlRoot}
}
\value{
list class with entryID,name,type,link, X of graphic,Y of graphic,width of graphic,height of graphic
}
\description{
Get KEGG pathway info from xmlRoot
}
\examples{
#getPathInfo(r)
}
