% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{getPValue}
\alias{getPValue}
\title{Calculate p-value and adjust p-value of KEGG pathway on input sample}
\usage{
getPValue(thispathwayID_DF, thisKEGGSpecies, thisInputSampleKEGG, N, n)
}
\arguments{
\item{thispathwayID_DF}{KEGG pathway ID}

\item{thisKEGGSpecies}{species in KEGG format}

\item{thisInputSampleKEGG}{input sample with in KEGG gene ID}

\item{N}{the number of all genes with KEGG gene ID}

\item{n}{the number of sample genes with KEGG gene ID}
}
\value{
vector in dataframe class containing p-value, adjust p-value,
N:the number of all genes with KEGG gene ID,
n the number of sample genes with KEGG gene ID,
M:the number of all genes in specific KEGG pathway,
m: the number of sample genes in specific KEGG pathway
}
\description{
Calculate p-value and adjust p-value of KEGG pathway on input sample
}
\examples{
getPValue("mmu00053","mmu",thisInputSampleKEGG,N,n)

}
