% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{getTotalPathNames}
\alias{getTotalPathNames}
\title{Get all KEGG pathway name for species}
\usage{
getTotalPathNames(KEGGSpecies)
}
\arguments{
\item{KEGGSpecies}{species in KEGG format}
}
\value{
all KEGG pathway name of input KEGG species
}
\description{
Get all KEGG pathway name for species
}
\examples{
getTotalPathNames("mmu")

}
