% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Key2GOEnrich.R
\name{GO2Enrich}
\alias{GO2Enrich}
\title{Perform Gene Ontology enrichment analysis}
\usage{
GO2Enrich(inputSample, inputSpecies, adjustMethod, filterValue, GOType,
  imgWidth, imgHeight)
}
\arguments{
\item{inputSample}{formatted input sample}

\item{inputSpecies}{human, mouse, rat}

\item{adjustMethod}{p.adjust.methods,
c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{filterValue}{must by numeric, value from 0 to 1}

\item{GOType}{BP,MF,CC}

\item{imgWidth}{the width of export file}

\item{imgHeight}{the height of export file}
}
\value{
data in dataframe class with pValue, adjust pValue of significate Gene Ontology
}
\description{
Perform Gene Ontology enrichment analysis
}
\examples{
egoDF<-GO2Enrich(inputSample=inputSample,
inputSpecies="mouse",
adjustMethod="fdr",
filterValue=0.05,
GOType="BP",
imgWidth=15,
imgHeight=20)
}
