% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs}
\alias{gs}
\title{Sample in GSEABase class}
\format{Sample in GSEABase class:
\describe{
  \item{geneIds}{entrez gene ID}
}}
\usage{
gs
}
\description{
The data from Affymetrix microarray of western diet feed mouse (WD)
vs. low fat diet feed mouse (LF)
were used as an example.
differential expressed genes were identified with
cutoff FDR<0.05 using limma R package.
}
\keyword{datasets}
