% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputSample}
\alias{inputSample}
\title{Data from sample file}
\format{A data frame with 721 rows and 10 variables:
\describe{
  \item{entrezgene}{entrez gene ID}
  \item{mgi_symbol}{gene symbol}
  \item{AveExpr}{average expression value of gene}
  \item{t}{t value}
  \item{P.Value}{p value}
  \item{adj.P.Val}{adjusted p value}
  \item{B}{B value}
  \item{log2FC}{log fold change value}
  \item{ensembl_gene_id}{ensembl gene ID}
  \item{KEGG_GeneID}{KEGG gene ID}
}}
\usage{
inputSample
}
\description{
The data from Affymetrix microarray of western diet feed mouse (WD)
vs. low fat diet feed mouse (LF) were used as an example.
748 differential expressed genes were identified with cutoff FDR<0.05 using limma R package.
}
\keyword{datasets}
