% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Key2KEGGEnrich.R
\name{KEGGEnricherMatrixRealTime}
\alias{KEGGEnricherMatrixRealTime}
\title{Get enrichment matrix of enriched KEGG pathway}
\usage{
KEGGEnricherMatrixRealTime(inputSample, inputSpecies, adjustMethod,
  filterMethod, filterValue)
}
\arguments{
\item{inputSample}{input formatted sample}

\item{inputSpecies}{human, mouse, rat}

\item{adjustMethod}{p.adjust.methods,
c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{filterMethod}{pValue or other character vector}

\item{filterValue}{must by numeric, value from 0 to 1}
}
\value{
data in dataframe class with pValue, adjust pValue of significate KEGG pathway
}
\description{
Get enrichment matrix of enriched KEGG pathway
}
\examples{
#thisKEGGEnricherMatrix<-KEGGEnricherMatrixRealTime(inputSample,"mouse","fdr","fdr",0.05)
}
