% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{KEGGID2EntrezID}
\alias{KEGGID2EntrezID}
\title{Get all KEGG gene ID and entrezgene ID in a specific species}
\usage{
KEGGID2EntrezID(KEGGSpecies)
}
\arguments{
\item{KEGGSpecies}{species in KEGG format}
}
\value{
all KEGG Gene ID and its entrezgene ID of input KEGG species
}
\description{
Get all KEGG gene ID and entrezgene ID in a specific species
}
\examples{
KEGGID2EntrezID("mmu")

}
