% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Key2KEGGEnrich.R
\name{mappingOnSpecifiedKEGGPathway}
\alias{mappingOnSpecifiedKEGGPathway}
\title{Download KEGG pathway files and mapping genes on pathway}
\usage{
mappingOnSpecifiedKEGGPathway(plotPathID, inputSampleKEGG, inputSpecies)
}
\arguments{
\item{plotPathID}{KEGG pathway ID}

\item{inputSampleKEGG}{input formatted sample}

\item{inputSpecies}{human, mouse, rat}
}
\value{
NA
}
\description{
Download KEGG pathway files and mapping genes on pathway
}
\examples{
mappingOnSpecifiedKEGGPathway("mmu05160",inputSample,"mouse")
}
