% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlot.R
\name{multiplot}
\alias{multiplot}
\title{Plot two image on same panel}
\usage{
multiplot(p1, p2)
}
\arguments{
\item{p1}{ggplot}

\item{p2}{ggplot}
}
\value{
plot
}
\description{
Plot two image on same panel
}
\examples{
multiplot(p1,p2)
}
