% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKGML.R
\name{parseEntry}
\alias{parseEntry}
\title{parseEntry}
\usage{
parseEntry(entry)
}
\arguments{
\item{entry}{this entry}
}
\value{
a list of node value
}
\description{
parseEntry
}
\examples{
path<-system.file(package = "Key2Enrich")
xmlPath<-paste(path,"/extdata",sep="")
xmlFile<-paste(xmlPath,sep="","/hsa04012.xml")
r<-parseKGMLFile(xmlFile)
kegg.nodes <- lapply(r[childIsEntry(r)], parseEntry)
\donttest{
nodeDF<-data.frame(KEGG_GeneID="NA",type=NA,link=NA,graphicName=NA,
                  graphicType=NA,graphicX=NA,graphicY=NA,
                  graphicWidth=NA,graphicHeight=NA)
nodeDF<-parseList2Dataframe(kegg.nodes,nodeDF)
head(nodeDF)
}
}
