% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKGML.R
\name{parseGraphic}
\alias{parseGraphic}
\title{Get info from list class of graphic info}
\usage{
parseGraphic(thisGeneKEGGID)
}
\arguments{
\item{thisGeneKEGGID}{gene KEGG ID}
}
\value{
character vector with name,type, X of graphic,Y of graphic,width of graphic,height of graphic
}
\description{
Get info from list class of graphic info
}
