% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKGML.R
\name{parseKGMLFile}
\alias{parseKGMLFile}
\title{Read XML file}
\usage{
parseKGMLFile(file)
}
\arguments{
\item{file}{file name}
}
\value{
xmlRoot of file
}
\description{
Read XML file
}
\examples{
path<-system.file(package = "Key2Enrich")
xmlPath<-paste(path,"/extdata",sep="")
xmlFile<-paste(xmlPath,sep="","/hsa04012.xml")
parseKGMLFile(xmlFile)
}
