% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKGML.R
\name{parseList2Dataframe}
\alias{parseList2Dataframe}
\title{Convert list class of node info to dataframe}
\usage{
parseList2Dataframe(nodeList, nodeDF)
}
\arguments{
\item{nodeList}{list class of node info}

\item{nodeDF}{output dataframe}
}
\value{
dataframe class of node info
}
\description{
Convert list class of node info to dataframe
}
\examples{
library(Key2Enrich)
path<-system.file(package = "Key2Enrich")
xmlPath<-paste(path,"/extdata",sep="")
xmlFile<-paste(xmlPath,sep="","/hsa04012.xml")
r<-parseKGMLFile(xmlFile)
kegg.nodes <- lapply(r[childIsEntry(r)], parseEntry)
nodeDF<-data.frame(KEGG_GeneID="NA",type=NA,link=NA,graphicName=NA,
                  graphicType=NA,graphicX=NA,graphicY=NA,
                  graphicWidth=NA,graphicHeight=NA)
nodeDF<-parseList2Dataframe(kegg.nodes,nodeDF)
head(nodeDF)
}
