% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{plotGradient}
\alias{plotGradient}
\title{Plot heatmap color key for node}
\usage{
plotGradient(thisWidth, thisHeight)
}
\arguments{
\item{thisWidth}{the width of color key}

\item{thisHeight}{the height of color key}
}
\value{
NA
}
\description{
Plot heatmap color key for node
}
\examples{
 plotGradient(120,5)

}
