% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{plotSigImg}
\alias{plotSigImg}
\title{Plot heat map rectangle on the KEGG pathway node}
\usage{
plotSigImg(filterPValueName, inputSampleKEGG, inputSpecies)
}
\arguments{
\item{filterPValueName}{p-value and adjust-pvalue matrix of sample on all KEGG pathways}

\item{inputSampleKEGG}{input sample with in KEGG gene ID}

\item{inputSpecies}{KEGGSpecies species in KEGG format}
}
\value{
NA
}
\description{
Plot heat map rectangle on the KEGG pathway node
}
