% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Key2ReactomeEnrich.R
\name{Reactome2Enrich}
\alias{Reactome2Enrich}
\title{Get enrichment matrix of enriched Reactome pathway}
\usage{
Reactome2Enrich(inputSample, inputSpecies, adjustMethod, filterValue, imgWidth,
  imgHeight)
}
\arguments{
\item{inputSample}{formatted input sample}

\item{inputSpecies}{human, mouse, rat}

\item{adjustMethod}{p.adjust.methods, c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{filterValue}{must by numeric, value from 0 to 1}

\item{imgWidth}{the width of export file}

\item{imgHeight}{the height of export file}

\item{filterMethod}{pValue or other character vector}
}
\value{
data in dataframe class with pValue, adjust pValue of significate Reactome pathway
}
\description{
Get enrichment matrix of enriched Reactome pathway
}
\examples{
reactomePValueMatrix<-Reactome2Enrich(inputSample,"mouse","fdr",0.05,15,20)
}
