% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{readInputFile}
\alias{readInputFile}
\title{Read samples files with different annotation ID}
\usage{
readInputFile(fileName, IDColumn, logFCColumn, IDType)
}
\arguments{
\item{fileName}{input file in csv format}

\item{IDColumn}{must be numberic,the column number of sample ID}

\item{IDType}{Human: entrezgene, hgnc_symbol, ensembl_gene_id, Mouse: entrezgene, mgi_symbol, ensembl_gene_id, Rat: entrezgene, rgd_symbol, ensembl_gene_id}
}
\value{
Sample file in data.frame format
}
\description{
Read samples files with different annotation ID
}
\examples{
entrezgene_sample<-readInputFile("sampleFile_entrezgene.csv",1,2,"entrezgene")

}
