% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{speciesConvert2Biomart}
\alias{speciesConvert2Biomart}
\title{Convert input species to biomart species}
\usage{
speciesConvert2Biomart(inputSpecies)
}
\arguments{
\item{inputSpecies}{human, mouse, rat}
}
\value{
species of biomart format,
human:hsapiens_gene_ensembl,
mouse:mmusculus_gene_ensembl,
rat:rnorvegicus_gene_ensembl
}
\description{
Convert input species to biomart species
}
\examples{
biomartSpecies<-speciesConvert2Biomart("mouse")

}
