% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{speciesKEGGConvert}
\alias{speciesKEGGConvert}
\title{Convert input species to KEGG species}
\usage{
speciesKEGGConvert(inputSpecies)
}
\arguments{
\item{inputSpecies}{human, mouse, rat}
}
\value{
species of KEGG format, human:hsa, mouse:mmu, rat:rno
}
\description{
Convert input species to KEGG species
}
\examples{
#KEGGSpecies<-speciesKEGGConvert("mouse")

}
