% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{speciesKEGGFlagConvert}
\alias{speciesKEGGFlagConvert}
\title{Convert input species to KEGG species with flag}
\usage{
speciesKEGGFlagConvert(inputSpecies)
}
\arguments{
\item{inputSpecies}{human, mouse, rat}
}
\value{
species of KEGG format with flag,
human:- Homo sapiens,
mouse:- Mus musculus,
rat:- Rattus norvegicus
}
\description{
Convert input species to KEGG species with flag
}
\examples{
#KEGGSpeciesFlag<-speciesKEGGFlagConvert("mouse")

}
