% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{textStamp}
\alias{textStamp}
\title{Add text stamp on image}
\usage{
textStamp(thisWidth)
}
\arguments{
\item{thisWidth}{the width of text stamp}
}
\value{
NA
}
\description{
Add text stamp on image
}
\examples{
# textStamp(100)

}
