% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{formatInputSample}
\alias{formatInputSample}
\title{Format input file to dataframe class with entrez ID, symbol and ensembl ID}
\usage{
formatInputSample(filename, IDColumn, logFCColumn, IDType, inputSpecies)
}
\arguments{
\item{filename}{input file in csv format}

\item{IDColumn}{must be numberic,the column number of sample ID}

\item{logFCColumn}{must be numberic,the column number of sample log fold change}

\item{IDType}{Human: entrezgene, hgnc_symbol, ensembl_gene_id, Mouse: entrezgene, mgi_symbol, ensembl_gene_id, Rat: entrezgene, rgd_symbol, ensembl_gene_id}

\item{inputSpecies}{human, mouse, rat}
}
\value{
input sample in dataframe class with entrez ID, symbol and ensembl ID
}
\description{
Format input file to dataframe class with entrez ID, symbol and ensembl ID
}
\examples{
inputSample<-formatInputSample("sampleFile_entrezgene.csv",1,2,"entrezgene","mouse")
}
