% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{getIndex}
\alias{getIndex}
\title{Get index of value in array for right color key}
\usage{
getIndex(thisArray, thisValue)
}
\arguments{
\item{thisArray}{color array}

\item{thisValue}{log fold change value}
}
\value{
index
}
\description{
Get index of value in array for right color key
}
\examples{
 #thisArray<-seq(0,2,length.out = 10)
#getIndex(thisArray,1.6)
}
