% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlot.R
\name{Key2EnrichBarplot}
\alias{Key2EnrichBarplot}
\title{Plot barplot}
\usage{
Key2EnrichBarplot(filterValuePathNameDF, type, imgWidth, imgHeight)
}
\arguments{
\item{filterValuePathNameDF}{matrix of enriched dataset}

\item{type}{the type of mapping database, KEGG, Reactome, BP, MF, CC}

\item{imgWidth}{the width of export file}

\item{imgHeight}{the height of export file}
}
\value{
data in dataframe class with pValue, adjust pValue
}
\description{
Plot barplot
}
\examples{
data(inputSample)
inputSample<-as.data.frame(inputSample)
\donttest{
thisKEGGEnricherMatrix<-KEGGEnricherMatrixRealTime(inputSample,
                                                   "mouse",
                                                   "fdr",
                                                   "fdr",
                                                   0.05)
Key2EnrichBarplot(thisKEGGEnricherMatrix,"KEGG",15,20)
}
}
