% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Key2KEGGEnrich.R
\name{Key2KEGGEnrich}
\alias{Key2KEGGEnrich}
\title{Performe enrichment analysis based on KEGG pathway,
plot barplot of significantly enriched KEGG pathway,
mapping genes on KEGG pathway image}
\usage{
Key2KEGGEnrich(inputSample, inputSpecies, adjustMethod, filterMethod,
  filterValue, ifbarplot, ifMapOnPath, type, imgWidth, imgHeight)
}
\arguments{
\item{inputSample}{input formatted sample}

\item{inputSpecies}{human, mouse, rat}

\item{adjustMethod}{p.adjust.methods,
c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{filterMethod}{pValue or other character vector}

\item{filterValue}{must by numeric, value from 0 to 1}

\item{ifbarplot}{barplot or not, TRUE/FALSE}

\item{ifMapOnPath}{pathway mapping or not, TRUE/FALSE}

\item{type}{KEGG, Reactome, BP,MF,CC}

\item{imgWidth}{the width of export file}

\item{imgHeight}{the height of export file}
}
\value{
data in dataframe class with pValue, adjust pValue of significate KEGG pathway
}
\description{
Performe enrichment analysis based on KEGG pathway,
plot barplot of significantly enriched KEGG pathway,
mapping genes on KEGG pathway image
}
\examples{
library(Key2Enrich)
data(inputSample)
inputSample<-as.data.frame(inputSample)
head(inputSample)
\dontrun{
KEGGSigMx<-Key2KEGGEnrich(inputSample,"mouse","fdr","fdr",0.05,TRUE,TRUE,"KEGG",15,20)
}
}
