% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKGML.R
\name{parseList2Dataframe}
\alias{parseList2Dataframe}
\title{Convert list class of node info to dataframe}
\usage{
parseList2Dataframe(nodeList, nodeDF)
}
\arguments{
\item{nodeList}{list class of node info}

\item{nodeDF}{output dataframe}
}
\value{
dataframe class of node info
}
\description{
Convert list class of node info to dataframe
}
\examples{
#parseList2Dataframe(nodeList,nodeDF)
}
