% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{readInputFile}
\alias{readInputFile}
\title{Read samples files with different annotation ID}
\usage{
readInputFile(fileName, IDColumn, logFCColumn, IDType)
}
\arguments{
\item{fileName}{input file in csv format}

\item{IDColumn}{must be numberic,the column number of sample ID}

\item{logFCColumn}{must be numberic,the column number of sample log fold change}

\item{IDType}{Human: entrezgene, hgnc_symbol, ensembl_gene_id,
Mouse: entrezgene, mgi_symbol, ensembl_gene_id,
Rat: entrezgene, rgd_symbol, ensembl_gene_id}
}
\value{
Sample file in data.frame format
}
\description{
Read samples files with different annotation ID
}
\examples{
path<-system.file(package = "Key2Enrich")
filePath<-paste(path,"/extdata",sep="")
file<-paste(filePath,sep="","/sampleFile.csv")
entrezgene_sample<-readInputFile(file,IDColumn=1,
                                 logFCColumn=2,IDType="mgi_symbol")
head(entrezgene_sample)

}
